Split `Blocks.tsx` into multiple files
